Move start to scaffold start and pass the bitcore-node configuration.
