+++ lib/scaffold/create.js
@@ -16,7 +16,8 @@ var BASE_CONFIG = {
     'address'
   ],
   datadir: './data',
-  network: 'livenet'
+  network: 'livenet',
+  port: 3001
 };
 
 var version;
